made a new file called videos_within_time_limit
This file gets the videos from the YouTube home page, YouTube search bar, or the YouTube sidebar and calculates the time of each video in seconds. Given a hard coded time_limit, the file gets rid of each video in the array that is longer than the time limit set.